Fix for relative path include problem introduced with my last PR: move the two new classes from PWGGAUtils/ to GammaConvBase/
