+++ PWGGA/GammaConvBase/AliConvEventCuts.h
@@ -22,7 +22,6 @@
 #include "AliVCaloTrigger.h"
 #include "AliTimeRangeCut.h"
 #include "AliEventCuts.h"
-#include "../PWGGAUtils/utils_TH1.h"
 
 class AliESDEvent;
 class AliAODEvent;
@@ -905,12 +904,13 @@ class AliConvEventCuts : public AliAnalysisCuts {
       bool                           fMapPtWeightsIsFilledAndSane;        ///<  flag to indicate if fMapPtWeightsAccessObjects is filled and sane
       TH2D                          *fHistoRelDiffNewOldMesonWeights_Pi0;
       TH2D                          *fHistoRelDiffNewOldMesonWeights_Eta;
-    //   utils_TH1                      fUtils_TH1;
-      Int_t                       fDebugLevel;                            ///< debug level for interactive debugging
+      
+      // keep this one last
+      Int_t                          fDebugLevel;                            ///< debug level for interactive debugging
   private:
 
       /// \cond CLASSIMP
-      ClassDef(AliConvEventCuts,97)
+      ClassDef(AliConvEventCuts,98)
       /// \endcond
 };
