+++ PWGGA/GammaConvBase/GammaConvUtilsTH1.cxx
@@ -1,4 +1,4 @@
-#include "utils_TH1.h"
+#include "GammaConvUtilsTH1.h"
 #include "TF1.h"
 #include "TH1.h"
