+++ PWGGA/GammaConvBase/PWGGAGammaConvBaseLinkDef.h
@@ -42,4 +42,7 @@
 #pragma link C++ class AliDalitzData+;
 #pragma link C++ class AliCaloSigmaCuts+;
 
+#pragma link C++ class utils_TH1+ ;
+#pragma link C++ class TH1_ExponentialInterpolation+ ;
+
 #endif
